+++ packages/rspress-plugin-file-tree/components/presets.ts
@@ -1,5 +1,0 @@
-export const presetClassName = 'rp-file-tree';
-
-export function buildClassName(fragment?: string) {
-  return fragment?.length ? `${presetClassName}-${fragment}` : presetClassName;
-}