+++ packages/rspress-plugin-live2d/components/Live2DWidget.d.ts
@@ -1,4 +1,0 @@
-import React from 'react';
-import type { Live2DWidgetProps } from '../src/typings';
-declare const Live2DWidget: React.FC<Live2DWidgetProps>;
-export default Live2DWidget;