+++ packages/rspress-plugin-supersub/src/index.ts
@@ -4,7 +4,7 @@ import {
   type RemarkParseSuperSubScriptOptions,
 } from './remark-plugins/parse-super-sub-script';
 
-import type { RspressPlugin } from '@rspress/shared';
+import type { RspressPlugin } from '@rspress/core';
 
 export interface RspressPluginSupersubOptions
   extends RemarkParseSuperSubScriptOptions {}
@@ -14,9 +14,7 @@ export default function rspressPluginSupersub(
 ): RspressPlugin {
   return {
     name: 'rspress-plugin-supersub',
-    config(config) {
-      return new PresetConfigMutator(config).disableMdxRs().toConfig();
-    },
+    // config not needed for Rspress V2
     markdown: {
       remarkPlugins: [
         [remarkParseSuperSubScript, <RemarkParseSuperSubScriptOptions>options],