+++ packages/rspress-plugin-back-to-top/CHANGELOG.md
@@ -1,5 +1,16 @@
 # rspress-plugin-back-to-top
 
+## 1.0.0
+
+### Major Changes
+
+- 732b475: feat: Compatible with V2 Rspress
+
+### Patch Changes
+
+- Updated dependencies [732b475]
+  - rspress-plugin-devkit@1.0.0
+
 ## 0.3.0
 
 ### Minor Changes