+++ packages/rspress-plugin-devkit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rspress-plugin-devkit
 
+## 1.0.0
+
+### Major Changes
+
+- 732b475: feat: Compatible with V2 Rspress
+
 ## 0.3.0
 
 ### Minor Changes